+++ docs/changes.rst
@@ -13,6 +13,7 @@ Not yet released.
 * Improved error handling on string removal.
 * Weblate now requires Python 3.7 or newer.
 * Fixed some write operations with projec token authentication.
+* Fixed string state tracking when the strings changes in the repository.
 
 Weblate 4.10.1
 --------------